fix: Navbar resize bug
The nav links disappear whenever the window is resize. This is due to the constant checking if the view width is less than 768px. I fix it by depending the state of the nav links to the hamburger menu. By doing this way, I can still display it to the larger screen size without modifying the display visibility of the nav links